% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_cSmoothEM.R
\name{compute_log_joint_observed_csmooth}
\alias{compute_log_joint_observed_csmooth}
\title{Penalized observed-data objective for csmooth_em (optional)}
\usage{
compute_log_joint_observed_csmooth(
  X,
  params,
  Q_K,
  lambda_vec,
  modelName,
  relative_lambda,
  eigen_tol = NULL,
  rw_q = 0L
)
}
\description{
Penalized observed-data objective for csmooth_em (optional)
}
\keyword{internal}
