% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_cSmoothEM.R
\name{compute_penalized_ELBO_csmooth}
\alias{compute_penalized_ELBO_csmooth}
\title{Penalized ELBO for csmooth_em (optional)}
\usage{
compute_penalized_ELBO_csmooth(
  X,
  Gamma,
  params,
  Q_K,
  lambda_vec,
  modelName,
  relative_lambda,
  eigen_tol = NULL,
  rw_q = 0L
)
}
\description{
Penalized ELBO for csmooth_em (optional)
}
\keyword{internal}
