% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08_SoftSmoothEM.R
\name{do_csmoothEM_ml_collapsed_weighted}
\alias{do_csmoothEM_ml_collapsed_weighted}
\title{Collapsed-ML csmoothEM with weighted E-step and weighted hyper-step}
\usage{
do_csmoothEM_ml_collapsed_weighted(
  object,
  data,
  feature_weights,
  iter = 1,
  record = TRUE,
  lambda_min = NULL,
  lambda_max = NULL,
  sigma_update = c("ml", "mstep"),
  verbose = FALSE
)
}
\description{
Collapsed-ML csmoothEM with weighted E-step and weighted hyper-step
}
\keyword{internal}
