% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_cSmoothEM.R
\name{.compute_Qbase_j}
\alias{.compute_Qbase_j}
\title{Internal: compute Q_base_j for coordinate j under relative_lambda}
\usage{
.compute_Qbase_j(pi, sigma2, Q_K, relative_lambda, modelName, j)
}
\description{
Internal: compute Q_base_j for coordinate j under relative_lambda
}
\keyword{internal}
