% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_cSmoothEM.R
\name{.refresh_map_mu_csmooth}
\alias{.refresh_map_mu_csmooth}
\title{Refresh pi/mu to the MAP solution under fixed sigma2 and lambda}
\usage{
.refresh_map_mu_csmooth(
  X,
  Gamma,
  params,
  Q_K,
  lambda_vec,
  modelName,
  relative_lambda
)
}
\description{
Refresh pi/mu to the MAP solution under fixed sigma2 and lambda
}
\keyword{internal}
