% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_cSmoothEM.R
\name{ESTEP_csmooth}
\alias{ESTEP_csmooth}
\title{Internal E-step for csmooth_em}
\usage{
ESTEP_csmooth(X, params, modelName)
}
\arguments{
\item{X}{n-by-d data matrix}

\item{params}{csmooth params list (pi, mu, sigma2)}

\item{modelName}{"homoskedastic" or "heteroskedastic"}
}
\value{
n-by-K responsibility matrix
}
\description{
Internal E-step for csmooth_em
}
\keyword{internal}
