% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_initialization.R
\name{initialize_ordering}
\alias{initialize_ordering}
\title{Initialize GMM parameters using an ordering method}
\usage{
initialize_ordering(
  X,
  K,
  method = c("PCA", "fiedler", "pcurve", "tSNE", "random"),
  discretization = c("equal", "quantile", "kmeans"),
  ...
)
}
\arguments{
\item{X}{numeric matrix (n x d).}

\item{K}{number of mixture components.}

\item{method}{ordering method.}

\item{discretization}{discretization method passed to make_init().}

\item{...}{forwarded to the ordering method function.}
}
\value{
list(pi, mu, sigma, keep_idx, cluster_rank, ordering)
}
\description{
Initialize GMM parameters using an ordering method
}
