% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_utils.R
\name{logsumexp}
\alias{logsumexp}
\title{Numerically stable log-sum-exp}
\usage{
logsumexp(x)
}
\arguments{
\item{x}{Numeric vector.}
}
\value{
log(sum(exp(x))) computed stably.
}
\description{
Numerically stable log-sum-exp
}
