% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_initialization.R
\name{make_default_init}
\alias{make_default_init}
\title{Default random initialization for a Gaussian mixture model}
\usage{
make_default_init(X, K, ordering = TRUE)
}
\arguments{
\item{X}{numeric matrix (n x d).}

\item{K}{number of mixture components.}

\item{ordering}{if TRUE, reorder components by a pivot dimension of mu.}
}
\value{
list(pi, mu, sigma)
}
\description{
Default random initialization for a Gaussian mixture model
}
