% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_initialization.R
\name{pcurve_ordering}
\alias{pcurve_ordering}
\title{Principal curve ordering}
\usage{
pcurve_ordering(
  X,
  smoother = c("smooth_spline", "lowess", "periodic_lowess"),
  thresh = 0.001,
  maxit = 10,
  stretch = 2,
  approx_points = FALSE,
  scale01 = TRUE,
  orient_by_pc1 = TRUE
)
}
\arguments{
\item{X}{numeric matrix (n x D).}

\item{smoother}{a function (recommended) or a name like "smooth_spline".}

\item{thresh, maxit, stretch, approx_points}{passed to princurve::principal_curve.}

\item{scale01}{logical; if TRUE, rescale returned t to [0,1].}

\item{orient_by_pc1}{logical; if TRUE, flip sign to align with PC1.}
}
\value{
list(t, keep_idx, fit)
}
\description{
Principal curve ordering
}
