% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_cSmoothEM.R
\name{print.csmooth_em}
\alias{print.csmooth_em}
\title{Print csmooth_em object}
\usage{
\method{print}{csmooth_em}(x, ...)
}
\arguments{
\item{x}{A \code{csmooth_em} object.}

\item{...}{Unused.}
}
\description{
Print csmooth_em object
}
