% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_mpcurve.R
\name{print.mpcurve}
\alias{print.mpcurve}
\title{Print an \code{mpcurve} fit}
\usage{
\method{print}{mpcurve}(x, ...)
}
\arguments{
\item{x}{An \code{mpcurve} object.}

\item{...}{Ignored.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print an \code{mpcurve} fit
}
