% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_methods.R
\name{print.smooth_em}
\alias{print.smooth_em}
\title{Print method for smooth_em objects}
\usage{
\method{print}{smooth_em}(x, ...)
}
\arguments{
\item{x}{A \code{smooth_em} object.}

\item{...}{Unused.}
}
\description{
Print method for smooth_em objects
}
