% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_cSmoothEM.R
\name{print.summary.csmooth_em}
\alias{print.summary.csmooth_em}
\title{Print summary of csmooth_em}
\usage{
\method{print}{summary.csmooth_em}(x, ...)
}
\arguments{
\item{x}{A \code{summary.csmooth_em} object.}

\item{...}{Unused.}
}
\description{
Print summary of csmooth_em
}
