% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_class.R
\name{print.summary.smooth_em}
\alias{print.summary.smooth_em}
\title{Print method for summary.smooth_em objects}
\usage{
\method{print}{summary.smooth_em}(x, ...)
}
\arguments{
\item{x}{A \code{summary.smooth_em} object.}

\item{...}{Unused.}
}
\description{
Print method for summary.smooth_em objects
}
