% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09_cavi.R
\name{summary.cavi}
\alias{summary.cavi}
\title{Summary method for \code{cavi} objects}
\usage{
\method{summary}{cavi}(object, ...)
}
\arguments{
\item{object}{A \code{cavi} object.}

\item{...}{Unused.}
}
\value{
An object of class \code{"summary.cavi"}.
}
\description{
Summary method for \code{cavi} objects
}
