% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_class.R
\name{summary.smooth_em}
\alias{summary.smooth_em}
\title{Summary method for smooth_em objects}
\usage{
\method{summary}{smooth_em}(object, ...)
}
\arguments{
\item{object}{A \code{smooth_em} object.}

\item{...}{Passed through (unused).}
}
\value{
An object of class \code{summary.smooth_em}.
}
\description{
Summary method for smooth_em objects
}
