% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_kriging.R
\name{compare_positions_by_history}
\alias{compare_positions_by_history}
\title{Compare posterior position summaries across two histories}
\usage{
compare_positions_by_history(
  res,
  h1,
  h2,
  type = c("mean", "max"),
  add_identity = TRUE,
  main = NULL
)
}
\description{
Compare posterior position summaries across two histories
}
