% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_kriging.R
\name{get_history_block}
\alias{get_history_block}
\title{Extract a (u, gamma) block for a given progressive history index}
\usage{
get_history_block(res, h, normalize_gamma = TRUE)
}
\description{
Extract a (u, gamma) block for a given progressive history index
}
