% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPCurver-package.R
\docType{package}
\name{MPCurver}
\alias{MPCurver-package}
\alias{MPCurver}
\title{MPCurver package}
\description{
CAVI-first trajectory and pseudotime inference with legacy smoothEM
backends retained for backward compatibility.
}
\author{
\strong{Maintainer}: Ziang Zhang \email{aguero.zhang@gmail.com}

}
\keyword{internal}
