% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_initialization.R
\name{PCA_ordering}
\alias{PCA_ordering}
\title{PCA ordering (PC1)}
\usage{
PCA_ordering(X, center = TRUE, scale = FALSE, scale01 = TRUE)
}
\arguments{
\item{X}{numeric matrix (n x D).}

\item{center}{logical.}

\item{scale}{logical.}

\item{scale01}{logical; if TRUE, rescale returned t to [0,1].}
}
\value{
list(t, keep_idx)
}
\description{
PCA ordering (PC1)
}
